Switched clpmn_all ints to long long ints

+++ scipy/special/special.h
@@ -111,27 +111,27 @@ void lpmn_all(T z, OutputMat1 res, OutputMat2 res_jac, OutputMat3 res_hess) {
 }
 
 template <typename T>
-std::complex<T> clpmn(long m, long n, long type, std::complex<T> z) {
+std::complex<T> clpmn(long long int m, long long int n, long long int type, std::complex<T> z) {
     return special::assoc_legendre_p(n, m, type, z);
 }
 
 template <typename T>
-void clpmn(long m, long n, long type, T z, T &res, T &res_jac) {
+void clpmn(long long int m, long long int n, long long int type, T z, T &res, T &res_jac) {
     special::assoc_legendre_p(n, m, type, z, res, res_jac);
 }
 
 template <typename T>
-void clpmn(long m, long n, long type, T z, T &res, T &res_jac, T &res_hess) {
+void clpmn(long long int m, long long int n, long long int type, T z, T &res, T &res_jac, T &res_hess) {
     special::assoc_legendre_p(n, m, type, z, res, res_jac, res_hess);
 }
 
 template <typename T, typename OutputMat1>
-void clpmn_all(long type, std::complex<T> z, OutputMat1 res) {
+void clpmn_all(long long int type, std::complex<T> z, OutputMat1 res) {
     special::assoc_legendre_p_all(type, z, res);
 }
 
 template <typename T, typename OutputMat1, typename OutputMat2>
-void clpmn_all(long type, std::complex<T> z, OutputMat1 res, OutputMat2 res_jac) {
+void clpmn_all(long long int type, std::complex<T> z, OutputMat1 res, OutputMat2 res_jac) {
     special::assoc_legendre_p_all(type, z, res, res_jac);
 }
